Connect search to backend (#55)
* Flex 1.16.1

* Implement FieldText

* Remove parents from FieldSelect and complete callerInformation tab

* add currentTarget value in case target value is null

* change FieldSelect store to field for consistency

* Extract RequiredAsterisk component

* Use FieldText and FieldSelect throughout the whole TabbedForms

* Use handleChange instead of handleCheckbox

* Update HrmFormPlugin to 0.3.1

* Initial commit for linting tools

* Complete configuring lint tools

* Initial config for plugin-hrm-form CI

* Restrict to Node10

* Trigger CI on push and PRs

* On push seems to be enough since it runs when pushing to a PR

* Set working-directory

* Set working-directory for each step

* Create fake appConfig.js

* Adds touched property to fieldType definition

* Increment plugin version

* Add step for creating temp files

* Add missing working directory

* Create /private folder

* Use craco directly to run the tests

* Adds name for checkout step

* Update test snapshots

* Update and remove unnecessary dependencies

* Increment plugin version

* Add Search tab

* Implement basic search form

* Fake search results

* Click name to show details

* Add tabs to SearchResultDetails

* Implement details for childInformation and caseInformation

* Fix hiding Caller Information

* Send a message to the caller when the counselor has ended the chat (#40)

* Update version (#41)

* Hover detail with max 200 chars on callSummary

* Click callSummary to open dialog with full text

* CallSummary hover details and dialog applied in SearchResultDetails

* Initial handleSelectSearchResult

* Add name and location objects to searchResults structure

* (No Review) CI working with flex-plugin (#42)

* Copy SearchResult values into TabbedForms

* Redirect to ChildInformation tab when selecting a SearchResult

* Make tabs scrollable

* Style SearchForm

* Remove Area and Time from SearchForm

* Code cleanup

* More code cleanup

* Increment plugin version

* Save counselorId and helpline

* Rename counselorId to twilioWorkerId

* Revert "Update version (#41)" (#45)

This reverts commit a1f62eca0854b630544eeb33463138e06bd9e4d1.

* Revert "Send a message to the caller when the counselor has ended the chat (#40)" (#46)

This reverts commit 04f129ceeebffe898e5a14308f20234cec8fbd88.

* Update plugin version

* Downgrade Flex to 1.14.1 (#48)

* Use new GET /contacts API and downgrade Flex to 1.14.1 (#47)

* Add CI status badge to README

* Add CI status badge to inner README

* Gian add duration (#51)

* Added comunication duration

* Changed ContactFormFactory.test

* Metadata added to form outside recursive form creation

* Added test for metadata inside form

* Required changes for PR

* Handled edge-cases where form is recreated

* Lint fix

* DRY

* Deleted old comment

* Connect to searchContacts backend

* Creates ConfigurationContext and withConfiguration HOC

* Remove localhost link

* Code review fixes

* Get rid of HrmFormController

* Increment plugin version
