+++ plugin-hrm-form/src/components/SearchResults.js
@@ -146,9 +146,9 @@ class SearchResults extends Component {
                     <StyledTableCell>{result.overview.callType}</StyledTableCell>
                     <StyledTableCell>{result.overview.categories}</StyledTableCell>
                     <StyledTableCell>{result.overview.counselor}</StyledTableCell>
-                    <StyledTableCell onClick={() => this.handleClickCallSummary(result.overview.callSummary)}>
-                      <Tooltip title={result.overview.callSummary.substr(0, 200)}>
-                        <span>{result.overview.callSummary}</span>
+                    <StyledTableCell onClick={() => this.handleClickCallSummary(result.overview.notes)}>
+                      <Tooltip title={result.overview.notes.substr(0, 200)}>
+                        <span>{result.overview.notes}</span>
                       </Tooltip>
                     </StyledTableCell>
                   </TableRow>
