+++ plugin-hrm-form/src/components/TaskView.js
@@ -1,9 +1,15 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 import { withTaskContext } from '@twilio/flex-ui';
 
 import HrmForm from './HrmForm';
 import { formType, taskType } from '../types';
+import { namespace, contactFormsBase } from '../states';
+import { Actions } from '../states/ContactState';
+import { handleBlur, handleCategoryToggle, handleFocus, handleSubmit } from '../states/ActionCreators';
+import { handleSelectSearchResult } from '../states/SearchContact';
 
 const wrapperStyle = {
   position: 'absolute',
@@ -24,9 +30,9 @@ class TaskView extends Component {
     thisTask: taskType.isRequired,
     form: formType.isRequired,
     handleBlur: PropTypes.func.isRequired,
-    handleCategoryToggle: PropTypes.func.isRequired,
     handleChange: PropTypes.func.isRequired,
     handleCallTypeButtonClick: PropTypes.func.isRequired,
+    handleCompleteTask: PropTypes.func.isRequired,
     handleSubmit: PropTypes.func.isRequired,
     handleFocus: PropTypes.func.isRequired,
     handleSelectSearchResult: PropTypes.func.isRequired,
@@ -41,7 +47,11 @@ class TaskView extends Component {
   }
 
   render() {
-    const { task, thisTask } = this.props;
+    const { task, thisTask, form } = this.props;
+
+    if (!task) {
+      return null;
+    }
 
     // If this task is not the active task, hide it
     const show = task && task.taskSid === thisTask.taskSid ? 'visible' : 'hidden';
@@ -49,12 +59,12 @@ class TaskView extends Component {
     return (
       <div style={{ ...wrapperStyle, visibility: show }}>
         <HrmForm
-          form={this.props.form}
-          handleBlur={this.props.handleBlur}
-          handleCategoryToggle={this.props.handleCategoryToggle}
+          form={form}
+          handleBlur={this.props.handleBlur(form, task.taskSid)}
+          handleCategoryToggle={handleCategoryToggle(form, this.props.handleChange)}
           handleChange={this.props.handleChange}
           handleCallTypeButtonClick={this.props.handleCallTypeButtonClick}
-          handleSubmit={this.props.handleSubmit}
+          handleSubmit={this.props.handleSubmit(form, this.props.handleCompleteTask)}
           handleFocus={this.props.handleFocus}
           handleSelectSearchResult={this.props.handleSelectSearchResult}
         />
@@ -63,4 +73,20 @@ class TaskView extends Component {
   }
 }
 
-export default withTaskContext(TaskView);
+const mapStateToProps = (state, ownProps) => {
+  // This should already have been created when beforeAcceptTask is fired
+  return {
+    form: state[namespace][contactFormsBase].tasks[ownProps.thisTask.taskSid],
+  };
+};
+
+const mapDispatchToProps = dispatch => ({
+  handleBlur: handleBlur(dispatch),
+  handleCallTypeButtonClick: bindActionCreators(Actions.handleCallTypeButtonClick, dispatch),
+  handleChange: bindActionCreators(Actions.handleChange, dispatch),
+  handleFocus: handleFocus(dispatch),
+  handleSubmit: handleSubmit(dispatch),
+  handleSelectSearchResult: bindActionCreators(handleSelectSearchResult, dispatch),
+});
+
+export default withTaskContext(connect(mapStateToProps, mapDispatchToProps)(TaskView));
